websocket/requirements.txt
appnope==0.1.2
attrs==21.2.0
backcall==0.2.0
bidict==0.21.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
click==8.0.3
cryptography==36.0.0
decorator==5.1.0
Flask==2.0.2
Flask-SocketIO==5.1.1
Flask-SQLAlchemy==2.5.1
gevent==21.8.0
greenlet==1.1.2
idna==3.3
ipython==7.30.0
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
jsonschema==4.2.1
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.23
ptyprocess==0.7.0
pycparser==2.21
Pygments==2.10.0
PyJWT==1.7.1
pyrsistent==0.18.0
python-engineio==4.3.0
python-socketio==5.5.0
requests==2.26.0
SQLAlchemy==1.4.27
traitlets==5.1.1
typing_extensions==4.0.0
urllib3==1.26.7
uWSGI==2.0.20
wcwidth==0.2.5
Werkzeug==2.0.2
zope.event==4.5.0
zope.interface==5.4.0
